#112: Moved useSearch hook to own file
